[79]: add banner for weekly book vip
