+++ common/uicomponent/src/main/AndroidManifest.xml
@@ -6,5 +6,6 @@
             <action android:name="android.support.customtabs.action.CustomTabsService" />
         </intent>
     </queries>
+    <uses-permission android:name="android.permission.INTERNET" />
 
 </manifest>
