+++ feature/archived/build.gradle.kts
@@ -61,4 +61,7 @@ dependencies {
 
     // Firebase
     implementation(libs.firebaseFirestore)
+
+    // Google Ad
+    implementation(libs.googleAd)
 }
