+++ feature/homepage/build.gradle.kts
@@ -60,4 +60,7 @@ dependencies {
 
     // Firebase
     implementation(libs.firebaseFirestore)
+
+    // Google Ad
+    implementation(libs.googleAd)
 }
